pyproject.toml
[project]
name = "platform-as-agent"
version = "0.1.0"
description = "Platform as Agent - an intelligent multi-agent system for the agent deployment lifecycle"
readme = "README.md"
license = "MIT-0"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk",
    "click>=8.0",
]

[project.optional-dependencies]
bedrock = [
    "boto3>=1.34",
]
strands = [
    "strands-agents>=0.1",
    "strands-agents-tools>=0.1",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pyyaml>=6.0",
    "ruff>=0.1",
]

[project.scripts]
plato = "platform_agent.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/platform_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
